+++ example/lib/example_renderer_maxlines.dart
@@ -1,0 +1,22 @@
+import 'package:flutter/material.dart';
+import 'package:light_html_editor/light_html_editor.dart';
+
+class RendererMaxLinesExample extends StatelessWidget {
+  const RendererMaxLinesExample({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      backgroundColor: Colors.grey,
+      body: Center(
+        child: SizedBox(
+          width: 400,
+          child: RichtextRenderer.fromRichtext(
+            "<b>Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet.</b>",
+            maxLines: 2,
+          ),
+        ),
+      ),
+    );
+  }
+}